Extended test cases for ProjectDAO, now covering all functionalities.

+++ MobileTimeRecordingTests/ProjectDAOTests.swift
@@ -13,6 +13,7 @@ class ProjectDAOTests: XCTestCase
 {
     var projects: [Project] = []
     
+    
     override func setUp()
     {
         super.setUp()
@@ -27,7 +28,10 @@ class ProjectDAOTests: XCTestCase
     {
         for project in projects
         {
-            projectDAO.removeProject(project)
+            if let project = projectDAO.getProject(project.id)
+            {
+                projectDAO.removeProject(project)
+            }
         }
         projects.removeAll(keepCapacity: false)
         
@@ -37,13 +41,164 @@ class ProjectDAOTests: XCTestCase
     
     func testAddProject_Valid_Pass()
     {
-        var pass = true
         for project in projects
         {
             projectDAO.addProject(project)
-            pass = pass && projectDAO.getProject(project.id)!.id == project.id
+        }
+        
+        var pass = true
+        for project in projects
+        {
+            if let project = projectDAO.getProject(project.id)
+            {
+                pass = pass && projectDAO.getProject(project.id)!.id == project.id
+            }
+            else
+            {
+                pass = false
+            }
         }
 
+        XCTAssert(pass, "Pass")
+    }
+    
+    
+    func testAddProject_Valid_Fail()
+    {
+        projectDAO.addProject(projects[0])
+        projectDAO.addProject(projects[1])
+        
+        var pass = true
+        for project in projects
+        {
+            if let project = projectDAO.getProject(project.id)
+            {
+                pass = pass && projectDAO.getProject(project.id)!.id == project.id
+            }
+            else
+            {
+                pass = false
+            }
+        }
+        
+        XCTAssert(!pass, "Pass")
+    }
+
+    
+    func testGetProject_Valid_Pass()
+    {
+        projectDAO.addProject(projects[0])
+        
+        var pass = true
+        if let project = projectDAO.getProject(projects[0].id)
+        {
+            pass = pass && projectDAO.getProject(project.id)!.id == project.id
+        }
+        else
+        {
+            pass = false
+        }
+        
+        XCTAssert(pass, "Pass")
+    }
+    
+    
+    func testGetProject_Valid_Fail()
+    {
+        var pass = true
+        for project in projects
+        {
+            if let project = projectDAO.getProject(project.id)
+            {
+                pass = pass && projectDAO.getProject(project.id)!.id == project.id
+            }
+            else
+            {
+                pass = false
+            }
+        }
+        
+        XCTAssert(!pass, "Pass")
+    }
+    
+    
+    func testArchiveProject_Valid_Pass()
+    {
+        for project in projects
+        {
+            projectDAO.addProject(project)
+            projectDAO.archiveProject(project)
+        }
+        
+        var pass = true
+        for project in projects
+        {
+            if let project = projectDAO.getProject(project.id)
+            {
+                pass = pass && projectDAO.getProject(project.id)!.isArchived
+            }
+            else
+            {
+                pass = false
+            }
+        }
+        
+        XCTAssert(pass, "Pass")
+    }
+    
+    
+    func testArchiveProject_Valid_Fail()
+    {
+        projectDAO.addProject(projects[0])
+        projectDAO.archiveProject(projects[0])
+        
+        projectDAO.addProject(projects[1])
+        projectDAO.archiveProject(projects[1])
+        
+        var pass = true
+        for project in projects
+        {
+            if let project = projectDAO.getProject(project.id)
+            {
+                pass = pass && projectDAO.getProject(project.id)!.isArchived
+            }
+            else
+            {
+                pass = false
+            }
+        }
+        
+        XCTAssert(!pass, "Pass")
+    }
+    
+    
+    func testRemoveProject_Valid_Pass()
+    {
+        for project in projects
+        {
+            projectDAO.addProject(project)
+        }
+        
+        
+        for project in projects
+        {
+            projectDAO.removeProject(project)
+        }
+        
+        
+        var pass = true
+        for project in projects
+        {
+            if let project = projectDAO.getProject(project.id)
+            {
+                pass = false
+            }
+            else
+            {
+                pass = pass && true
+            }
+        }
+        
         XCTAssert(pass, "Pass")
     }
 }
